# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ovumcy API
  version: 0.9.5
  summary: HTTP API surface of the ovumcy-web self-hosted deployment.
  description: |
    This document specifies the **current** state of the JSON-emitting HTTP
    surface of `ovumcy-web`. It is the authoritative description of what the
    server actually serves today.

    Stability — per [CONTRIBUTING.md](../CONTRIBUTING.md#api-stability-contract):

    - `/api/v1/*` is the canonical, stable HTTP surface. Wrappers and external
      integrations should target this prefix exclusively.
    - Field additions are non-breaking and may ship in any minor release.
    - Field renames, removals, status code changes, route moves, and error key
      changes are breaking and require a new major version (`/api/v2/*`).
    - The export payload (`/api/v1/exports/*`) follows the separate stability
      contract documented in [docs/export.md](export.md).

    Content negotiation — every endpoint serves both JSON and HTML/HTMX from the
    same handler. This document covers **only** the JSON contract. Clients must
    send `Accept: application/json` to opt into the JSON shape. Without that
    header the server returns HTML or `303` redirects suitable for browser
    navigation, and the corresponding response is not part of this contract.

    Authentication & authorization — most endpoints require the sealed
    `ovumcy_auth` session cookie issued by `POST /api/v1/sessions`. Endpoints
...
    for session semantics.

...

...

...
  license:
...
servers:
...
security:
  - sessionCookie: []
    csrfToken: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ovumcy_auth
      description: |
        Sealed session cookie issued by `POST /api/v1/sessions` (or by completing
        the OIDC sign-in flow). Verified against `users.auth_session_version` on
        every request. See [SECURITY.md](../SECURITY.md#cookies).
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        Double-submit CSRF token validated against the `ovumcy_csrf` cookie.
        Browser clients usually submit it via the `csrf_token` form field
        instead of the header. Pure-GET, register/login/forgot-password
        request paths, and the `/auth/oidc/callback` provider form-post path
        are exempt. See [CSRF middleware configuration](../cmd/ovumcy/main.go).
...
  responses:
    Unauthorized:
      description: Missing or invalid session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { error: "unauthorized" }
...
    ValidationError:
      description: Request body or query parameters failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { error: "invalid input" }
    RateLimited:
      description: Rate limit exceeded.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { error: "too many attempts" }
...
  schemas:
    ApiError:
      type: object
      description: |
        Standard error envelope. The top-level `error` string is the legacy
        key kept for backward compatibility. The `error_detail` object carries
        the structured form for programmatic clients — `category` identifies
        the error class (validation, unauthorized, rate_limited, …) and
        `target` distinguishes global errors from form-level ones. Clients
        should branch on the HTTP status code plus `error_detail.category`
        rather than the key string, and treat unknown keys/categories as their
        generic status-class failure.
      required: [error, error_detail]
      properties:
        error:
          type: string
          example: "auth.invalid_credentials"
        error_detail:
          $ref: '#/components/schemas/ApiErrorDetail'
      additionalProperties: false
    ApiErrorDetail:
      type: object
      required: [key, category, target]
      properties:
        key:
          type: string
          description: Same value as the top-level `error` string. Duplicated here so a client that consumes only `error_detail` can branch without parsing the parent envelope.
          example: "auth.invalid_credentials"
        category:
          type: string
          enum: [validation, not_found, conflict, unauthorized, forbidden, rate_limited, internal]
          description: Error class. Stable enum — new values may be added in any release but existing values do not change meaning.
        target:
          type: string
          enum: [global, auth_form, settings_form]
          description: Where the error originated. `global` errors are unscoped; form targets indicate the error belongs inside an auth or settings form so HTMX-style clients can render it inline.
      additionalProperties: false
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
      additionalProperties: false
    OkRedirectResponse:
      type: object
      required: [ok, redirect]
      properties:
        ok:
          type: boolean
          enum: [true]
        redirect:
          type: string
          description: Suggested next URL path for the client to follow.
          example: "/auth/oidc/logout"
      additionalProperties: false
    RegisterRequest:
      type: object
      required: [email, password, confirm_password]
      properties:
        email:
          type: string
          format: email
          description: Account email. Normalized server-side (lowercased, trimmed).
        password:
          type: string
          format: password
          minLength: 8
          description: Must satisfy the password policy — 8+ Unicode code points, at least one uppercase letter, one lowercase letter, and one digit.
        confirm_password:
          type: string
          format: password
      additionalProperties: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        remember_me:
          type: boolean
          default: false
          description: When true, the issued `ovumcy_auth` cookie carries an explicit ~7-day Expires. Otherwise the cookie is browser-session-scoped.
      additionalProperties: false
    LoginTOTPRequiredResponse:
      type: object
      description: Returned when the password is correct and the account has TOTP 2FA enabled. The caller must follow up with `POST /api/v1/sessions/2fa-challenge`.
      required: [requires_totp]
      properties:
        requires_totp:
          type: boolean
          enum: [true]
      additionalProperties: false
    TOTPLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: "^[0-9]{6}$"
          description: 6-digit code from the user's TOTP authenticator.
      additionalProperties: false
    ForgotPasswordRequest:
      type: object
      required: [email, recovery_code]
      properties:
        email:
          type: string
          format: email
        recovery_code:
          type: string
          pattern: "^OVUM-[0-9A-F]{4}-[0-9A-F]{4}-[0-9A-F]{4}$"
          description: Single-use recovery code minted at registration or by `POST /api/v1/users/current/recovery-code`.
      additionalProperties: false
    ResetPasswordRequest:
      type: object
      required: [password, confirm_password]
      properties:
        password:
          type: string
          format: password
          minLength: 8
        confirm_password:
          type: string
          format: password
          minLength: 8
      additionalProperties: false
...
    CurrentUser:
      type: object
      required: [id, email, role, onboarding_completed, local_auth_enabled, must_change_password]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [owner] }
        onboarding_completed: { type: boolean }
        local_auth_enabled: { type: boolean }
        must_change_password: { type: boolean }
      additionalProperties: false
...
    PasswordOnlyRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          format: password
...
paths:
...
  /api/v1/users:
    post:
      tags: [auth]
      summary: Create a new owner account.
      description: |
        Returns the same response shape (a single sealed `ovumcy_register_pickup`
        cookie of fixed length, redirect to `/register/welcome`) for new and
        duplicate emails. See [SECURITY.md "Register enumeration"](../SECURITY.md#register-enumeration-pickup-cookie-follow-up-oracle-residual).
        Per-IP rate limit: default 8 / 15 min.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '303':
          description: Redirect to `/register/welcome` with the sealed pickup cookie. Shape is identical for new and duplicate emails.
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/sessions:
    post:
      tags: [auth]
      summary: Sign in with email and password.
      description: |
        Per-IP rate limit (default 8 / 15 min) plus per-account
        `AuthAttemptPolicy` budget. Constant-time password verification with
        equalized bcrypt timing on the missing-user branch.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: |
            Either signed in (issues `ovumcy_auth`) or 2FA challenge required.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - $ref: '#/components/schemas/LoginTOTPRequiredResponse'
              examples:
                signedIn:
                  summary: Signed in
                  value: { ok: true }
                totpRequired:
                  summary: 2FA challenge required
                  value: { requires_totp: true }
        '303':
          description: "When `Accept: application/json` is not set, redirect to `/dashboard`, `/auth/2fa`, or `/reset-password`."
        '401':
          description: Invalid credentials.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example: { error: "auth.invalid_credentials" }
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/sessions/2fa-challenge:
    post:
      tags: [auth]
      summary: Submit the TOTP code in the 2FA challenge.
      description: |
        Requires the `ovumcy_totp_pending` cookie issued by a successful
        password login. Per-account TOTP rate limit: 5 attempts / 15 min.
        Replayed codes within the 30-second step are rejected with the same
        response shape as an invalid code.
      security:
        - csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TOTPLoginRequest' }
      responses:
        '200':
          description: 2FA verified; `ovumcy_auth` issued.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '303':
          description: Browser redirect to `/`.
        '401':
          description: Invalid code or pending session expired.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/sessions/current:
    delete:
      tags: [auth]
      summary: Revoke current session.
      description: |
        Bumps `auth_session_version` (invalidating every other auth cookie for
        the account), clears auth-related cookies, and may return an OIDC
        end-session bridge URL if the session was OIDC-backed.
        Per-account rate limit: 60 / 15 min.
      responses:
        '200':
          description: Logged out.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/OkResponse'
                  - $ref: '#/components/schemas/OkRedirectResponse'
        '303':
          description: Browser redirect to `/login` or the OIDC logout bridge path.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/password-resets:
    post:
      tags: [auth]
      summary: Begin password reset via recovery code.
      description: |
        Validates the `(email, recovery_code)` pair without leaking which one
        was wrong. Per-IP rate limit: default 8 / 1 hour.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ForgotPasswordRequest' }
      responses:
        '303':
          description: Redirect to `/reset-password` with a sealed reset cookie.
        '401':
          description: Recovery code does not match the supplied email.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/password-resets/redeem:
    post:
      tags: [auth]
      summary: Set a new password using the reset cookie.
      description: |
        Requires the `ovumcy_reset_password` cookie issued by `POST
        /api/v1/password-resets`. Bumps `auth_session_version` on success.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResetPasswordRequest' }
      responses:
        '303':
          description: Redirect to `/login` after success.
        '422':
          $ref: '#/components/responses/ValidationError'

  # =========================================================================
  # onboarding
  # =========================================================================
...
  /api/v1/users/current:
    get:
      tags: [settings]
      summary: Identity of the session subject.
      description: |
        Returns the minimum representation of the signed-in user a wrapper
        needs to learn who it is talking to: identity (id, email, display
        name), authorization tier (role), and lifecycle flags it must respect
        before mutating calls (onboarding_completed, local_auth_enabled,
        must_change_password). Sensitive fields are never included.
      responses:
        '200':
          description: Current user shape.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      tags: [settings]
      summary: Delete the current account and all related data (owner only).
      description: |
        Requires the current password. Deletes daily_logs, all symptoms for the
        user (including built-ins), the users row, and cascade-deletes
        `oidc_identities`. See [SECURITY.md "Retention and Deletion"](../SECURITY.md#retention-and-deletion).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PasswordOnlyRequest' }
      responses:
        '200':
          description: Account deleted. Auth cookies cleared.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
...